+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Fixes:
 * Don't reload system-settings while user is editing [#99](https://github.com/O-X-L/ansible-webui/issues/99)
 * Updated inventory-list logic to allow for dynamic-inventories [#89](https://github.com/O-X-L/ansible-webui/issues/89)
 * Form-submit not working after a previous validation-error
+* Job-specific extra-vars not merged with execution-prompt-vars [#112](https://github.com/O-X-L/ansible-webui/issues/112)
 
 Project: